requirements.txt
PySide6
PyYaml
pyinstaller